Review

+++ onlineshop-app/app/Http/Controllers/ProductController.php
@@ -3,16 +3,17 @@
 namespace App\Http\Controllers;
 
 use App\Models\Product;
+use App\Models\Review;
+use Illuminate\Http\Request;
 use App\Http\Resources\ProductResource;
-use App\Http\Requests\ProductRequest;
-use Illuminate\Http\Response;
 
 class ProductController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth')->except(['index', 'show']);
+    }
 
-    /*public function __construct(){
-        $this->authorizeResource(Product::class);
-    }*/
     /**
      * Display a listing of the resource.
      *
@@ -20,9 +21,12 @@ class ProductController extends Controller
      */
     public function index()
     {
-        return ProductResource::collection(Product::all());
+        $products = Product::with('user:id,name')
+            ->withCount('reviews')
+            ->latest()
+            ->paginate(20);
+        return response()->json(['products' => $products]);
     }
-
 
     /**
      * Store a newly created resource in storage.
@@ -30,56 +34,75 @@ class ProductController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(ProductRequest $request)
+    public function store(Request $request)
     {
-        $created_product = Product::create($request->validated());
-        return new ProductResource($created_product);
+        $request->validate([
+            'name' => 'required|string',
+            'description' => 'required|string',
+            'price' => 'required|numeric|min:0',
+        ]);
+
+        $product = new Product;
+        $product->name = $request->name;
+        $product->description = $request->description;
+        $product->price = $request->price;
+
+        auth()->user()->products()->save($product);
+        return response()->json(['message' => 'Product Added', 'product' => $product]);
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
+     * @param  \App\Product  $product
      * @return \Illuminate\Http\Response
      */
-    public function show($id)
+    public function show(Product $product)
     {
-        return new ProductResource(Product::findOrFail($id));
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function edit($id)
-    {
-        //
+        $product->load(['reviews' => function ($query) {
+            $query->latest();
+        }, 'user']);
+        return response()->json(['product' => $product]);
     }
 
     /**
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
+     * @param  \App\Product  $product
      * @return \Illuminate\Http\Response
      */
-    public function update(ProductRequest $request, Product $product)
+    public function update(Request $request, Product $product)
     {
-        $product->update($request->validated());
-        return new ProductResource($product);
+        if (auth()->user()->id !== $product->user_id) {
+            return response()->json(['message' => 'Action Forbidden']);
+        }
+        $request->validate([
+            'name' => 'required|string',
+            'description' => 'required|string',
+            'price' => 'required|numeric',
+        ]);
+
+        $product->name = $request->name;
+        $product->description = $request->description;
+        $product->price = $request->price;
+        $product->save();
+
+        return response()->json(['message' => 'Product Updated', 'product' => $product]);
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param  \App\Product  $product
      * @return \Illuminate\Http\Response
      */
     public function destroy(Product $product)
     {
+        if (auth()->user()->id !== $product->user_id) {
+            return response()->json(['message' => 'Action Forbidden']);
+        }
         $product->delete();
-
+        return response()->json(null, 204);
     }
 }
